requirements.txt
Django==5.0.2
django-filter==23.5
django-rest-auth==0.9.5
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
load-dotenv==0.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
requests==2.31.0
django-jet-reboot==1.3.7
sentry-sdk==1.40.4
social-auth-app-django==5.4.0
social-auth-core==4.5.3
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.2.1
gunicorn==21.2.0

asgiref==3.7.2
celery==5.3.6
redis==5.0.1
django_celery_results=2.5.1
